Fix jQuery check
Trying to evaluate 'jQuery' when it is undefined would throw a
ReferenceError to the console. Trying to retrieve it on the global
`window` instead will result in it being `undefined` if not defined.

An alternative would be to compare the value of `typeof jQuery` to
'undefined', if we are concerned that `window` might not be the global
scope.