Add recipe for using with respond_to/format.js
Closes #53.

Update the README with a recipe for using the gem when `respond_to`/`format.js` blocks are used. By restricting the action to only answering AJAX requests, we restrict users accessing the partial by typing the action's path in the URL.

When `format.js` is used, Rails automatically sets the response's `Content-Type` to `text/javascript`. This causes the `$.ajax` call to incorrectly parse the response, failing to replace the content.

We need to explicitly set the response's `Content-Type` to `text/html` in the render call to fix that.