Add support for retry count header
Currently there's no way to tell on backend if request is being retried
or not. This change adds support for optional `retry_count_header`
parameter which configures request to include retry count in specified
header. (This is for example useful when you for example know that
previous request timed out and you want to reduce workload to try to
make it on time)

This change also fixes `retry_count` which is currently broken if
`retry_delay` is not specified.